Bugfix: Incorrect mode when featureless make is building settings.
This is a forward port of commit e70ac676f1907f17bcf24e428008183b8ccb96f5 from the 0.6 branch.

In the 0.6, this problem happens only on printing.
In this branch, the 0.7 development branch, this is used in different parts of the code and may have unintended functional consequences.

Oddly, this problem is not observed for the same files that the problem can be seen when using the 0.6 branch built programs.

I have opted to fix the problem rather than to investigate further.
The logic is clearly flawed and the fix is identical from the 0.6 branch.