Progress: Begin next development pass on fl_fss_payload_header_map().
Preserve the error status when exiting by checking and returning at the end of the function (before setting F_okay).

Fix the logic regarding the number macros when it comes to expanding an array of numbers into multiple destinations.
The object key must be set.
The macro needs to be accepting the destinations array rather than the specific destination string.
The case where f_fss_payload_header_map_flag_join_digits_e is set is not yet tested or reviewed.